% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsyllabus.R
\name{zsyllabus}
\alias{zsyllabus}
\title{My UMD Syllabus template}
\usage{
zsyllabus(...)
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document()}}.}
}
\value{
An R Markdown format object
}
\description{
A template to construct a syllabus formatted to follow the
University of Maryland Accessible Syllabus template. I tried my best to
make this template identical to the document shared by the Teaching and
Learning Transformation Center.

YAML header fields\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{title} \tab Title of the class \cr
   \code{subtitle} \tab (course number) \cr
   \code{term} \tab Season/Year \cr
   \code{author} \tab Name(s) of the professor(s) \cr
   \code{date} \tab start and end dates the course \cr
   \code{email} \tab Your email. \cr
   \code{officehours} \tab Your office hours. \cr
   \code{office} \tab Your office location. \cr
   \code{classroom} \tab The classroom location. \cr
   \code{classhours} \tab When/for how long the class meets in a given session. \cr
   \code{assistant} \tab Name(s) of teaching assistant(s) \cr
}
}
