% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarticle.R
\name{zarticle}
\alias{zarticle}
\title{Article template}
\usage{
zarticle(...)
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\value{
A custom article template
}
\description{
A template for academic articles. Custom made by Isaiah somehow
}
