% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarticle.R
\name{zarticle}
\alias{zarticle}
\title{Zarticle template}
\usage{
zarticle(...)
}
\arguments{
\item{...}{Arguments to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}.}
}
\description{
Zarticle template
}
