% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zsyllabus.R
\name{zsyllabus}
\alias{zsyllabus}
\title{My UMD Syllabus template}
\usage{
zsyllabus(
  logo = NULL,
  logo_height = "1.2cm",
  short_title = NULL,
  footer_on_first_page = FALSE,
  ...
)
}
\arguments{
\item{logo}{File path to Company logo file. The logo will be printed on
the top right corner of the first page and the size will be adjusted to
\code{logo_height}.}

\item{logo_height}{Height of the logo image. This logo image will be scaled
to height and the default height is 1.2cm.}

\item{short_title}{A shortened version of the course title or the course number to be printed
on the top-right corner of every page, except the front page}

\item{footer_on_first_page}{T/F value for whether the institution info footer
will be displayed on the front page. Default is FALSE}

\item{...}{other options to be passed to \code{\link[rmarkdown:pdf_document]{rmarkdown::pdf_document}}. See
?rmarkdown::pdf_document for details.}
}
\value{
An R Markdown format object
}
\description{
A template to construct a syllabus formatted to follow the
University of Maryland Accessible Syllabus template. I tried my best to
make this template identical to the document shared by the Teaching and
Learning Transformation Center.

YAML header fields\tabular{ll}{
   FIELD \tab DESCRIPTION \cr
   \code{title} \tab Title of the class \cr
   \code{subtitle} \tab (course number) \cr
   \code{term} \tab Season/Year \cr
   \code{author} \tab Name(s) of the professor(s) \cr
   \code{date} \tab start and end dates the course \cr
   \code{email} \tab Your email. \cr
   \code{logo} \tab logo from institution \cr
   \code{officehours} \tab Your office hours. \cr
   \code{office} \tab Your office location. \cr
   \code{classroom} \tab The classroom location. \cr
   \code{classhours} \tab When/for how long the class meets in a given session. \cr
   \code{assistant} \tab Name(s) of teaching assistant(s) \cr
}
}
